Recipeにshareカラム追加

+++ db/migrate/20230608060351_create_list_storages.rb
@@ -3,7 +3,7 @@ class CreateListStorages < ActiveRecord::Migration[6.1]
     create_table :list_storages do |t|
       t.integer :list_id,   null: false
       t.integer :recipe_id, null: false
-      t.float :quantity,   null: false
+      t.decimal :quantity,   null: false
 
       t.timestamps
     end
